Change EnvironmentImpl
