+++ src/test/kotlin/ru/hse/environment/SystemEnvironmentTest.kt
@@ -1,14 +1,13 @@
 package ru.hse.environment
 
 import org.junit.jupiter.api.Test
-import ru.hse.Environment
 import kotlin.test.assertEquals
 
 class SystemEnvironmentTest {
 
     @Test
     fun `test system environment works`() {
-        val env: Environment = TODO("Not yet implemented")
+        val env: Environment = EnvironmentImpl(null, System.getenv())
         val map = System.getenv()
         assertEquals(map["HOME"], env.get("HOME"))
         assertEquals(map.getOrDefault("AoAoA", ""), env.get("AoAoA"))
